fix lcm load speed
