+++ dimos/msgs/sensor_msgs/CameraInfo.py
@@ -20,7 +20,6 @@ import time
 from dimos_lcm.sensor_msgs import CameraInfo as LCMCameraInfo
 from dimos_lcm.std_msgs.Header import Header
 import numpy as np
-import rerun as rr
 
 # Import ROS types
 try:
@@ -405,6 +404,8 @@ class CameraInfo(Timestamped):
         Returns:
             rr.Pinhole archetype for logging to Rerun
         """
+        import rerun as rr
+
         # Extract intrinsics from K matrix
         # K = [fx, 0, cx, 0, fy, cy, 0, 0, 1]
         fx, fy = self.K[0], self.K[4]
